Fix: Resolve multiple issues including temperature settings, search quality, and UI bugs
- Fix Chairman behavior at temperature 0.0
- Fix duplicate ranking extraction
- Improve search quality (switch to general web search)
- Fix DuckDuckGo package warning
- Fix variable tag overlap in Settings UI
- Add provider badge to Chairman in Stage 3
- Fix provider labeling for Groq models
- Ensure settings persistence for temperatures
