+++ frontend/src/components/Settings.css
@@ -1093,7 +1093,12 @@ input:checked+.slider:before {
 .prompt-help {
   font-size: 12px;
   color: #9ca3af;
-  margin: 0;
+  margin: 0 0 8px 0;
+  display: flex;
+  flex-wrap: wrap;
+  align-items: center;
+  gap: 6px;
+  line-height: 1.6;
 }
 
 .prompt-help code {
